# Improve binary exporter perf