Fixes https://github.com/npgsql/npgsql/issues/4944

After this PR, sync exporting ends up being 10% faster on my machine (M1 Max) than using sync NpgsqlDataReader + SequentialAccess. Testing was done based on the project that was linked to here https://github.com/npgsql/npgsql/issues/4944#issuecomment-1441500430.

PR mostly consists of splitting apart sync and async paths. Having sync column reads run at the highest perf is as important here as it is in NpgsqlDataReader with GetFieldValue.