@NinoFloris 2024-03-17
```

BenchmarkDotNet v0.13.12, macOS Sonoma 14.2.1 (23C71) [Darwin 23.2.0]
Apple M1 Max, 1 CPU, 10 logical and 10 physical cores
.NET SDK 9.0.100-preview.1.24101.2
  [Host]     : .NET 8.0.0 (8.0.23.53103), Arm64 RyuJIT AdvSIMD
  DefaultJob : .NET 8.0.0 (8.0.23.53103), Arm64 RyuJIT AdvSIMD


```

With PR:

| Method           | Mean     | Error   | StdDev   | Ratio | RatioSD |
|----------------- |---------:|--------:|---------:|------:|--------:|
| BufferedReader   | 559.1 ms | 9.89 ms | 11.38 ms |  1.27 |    0.02 |
| SequentialReader | 493.1 ms | 9.53 ms |  8.91 ms |  1.12 |    0.02 |
| Export           | 442.1 ms | 2.45 ms |  2.29 ms |  1.00 |    0.00 |

Without: 

| Method           | Mean     | Error   | StdDev   | Ratio | RatioSD |
|----------------- |---------:|--------:|---------:|------:|--------:|
| BufferedReader   | 552.6 ms | 8.41 ms |  7.02 ms |  0.91 |    0.02 |
| SequentialReader | 496.3 ms | 9.78 ms | 10.47 ms |  0.82 |    0.02 |
| Export           | 605.1 ms | 6.14 ms |  5.44 ms |  1.00 |    0.00 |