@NinoFloris 2024-03-17
Just for testing I rebased this PR onto https://github.com/npgsql/npgsql/pull/5631 (which is based on https://github.com/npgsql/npgsql/pull/5480 and https://github.com/npgsql/npgsql/pull/5476) to get the following results


| Method           | Mean     | Error   | StdDev  | Ratio | RatioSD |
|----------------- |---------:|--------:|--------:|------:|--------:|
| BufferedReader   | 453.6 ms | 8.33 ms | 7.79 ms |  1.06 |    0.02 |
| SequentialReader | 458.5 ms | 7.95 ms | 7.04 ms |  1.07 |    0.02 |
| Export           | 428.4 ms | 2.47 ms | 2.31 ms |  1.00 |    0.00 |