@roji 2024-03-17
Out of curiosity, how many rows/columns are being exported here? Maybe post the benchmark source?

(not pushing back, just interested in more details)