@NinoFloris 2024-03-17
Benchmark is a bit hacked but it's basically https://github.com/npgsql/npgsql/issues/4944#issuecomment-1441500430

So that's 1 million rows being read