@NinoFloris 2024-03-17
https://github.com/SoftStoneDevelop/NpgsqlBinaryExporter/blob/main/Src/NpgsqlBenchmark/Extensions/Extension.cs#L124-L130

This pattern is why SkipIfNull/SkipNull would be useful, it would also show perf wins in datasets containing a lot of nulls. (here only the identification_id is null 50% of the time)