@NinoFloris 2024-03-28
Backported to 8.0.3 via 66969a5